backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-keycloak==3.7.0
casbin==1.36.2
casbin-sqlalchemy-adapter==1.4.0

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# HTTP client for external integrations
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Environment and configuration
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Property-based testing
hypothesis==6.92.1

# LLM Provider integrations
openai==1.3.7
anthropic==0.7.8
ollama==0.1.7

# Additional utilities for LLM integration
tiktoken==0.5.2
tenacity==8.2.3